Fixes issue where msguser is None
This issue occurs when a new user is added
to the slack team while slask is running.
Slask does not update its list of users and
when this new user generates an event slask
tries to lookup the user in its internal list
of users and gets None. None is never accounted
for in the code and it is always assumed that
the user will always be in the list.